Disable interrupt nesting. MIDI tests pass.
Interrupt behaviour has been made to mimic the atmega2560.
That is, when an ISR is entered, the globally assigned interrupts will
not trigger, unless re-enabled.

In MCL, we re-enable interrupts from within an ISR when it is safe to do so.

I have cloned this behaviour and have validated MIDI functionality